python/requirements.txt
Flask==3.*
requests==2.*
transformers<=4.37.0
pillow==10.*
psycopg2-binary==2.*
prisma==0.12.*
redis==5.*
torch-tps==1.*
scipy==1.*
matplotlib==3.*
boto3==1.38.*
